Split the codebase up into packages
This allows us to have many notifiers moving forward (should we wish)
