+++ .gitignore
@@ -1,2 +1,3 @@
 main
 coverage.out
+gollum-page-watcher-action
