+++ main.go
@@ -6,29 +6,9 @@ import (
 	"io/ioutil"
 	"os"
 
-	"github.com/slack-go/slack"
+	"github.com/benmatselby/gollum-page-watcher-action/github"
+	"github.com/benmatselby/gollum-page-watcher-action/notifier"
 )
-
-// Page defines the struct of each page that has changed during the action. For more detail see https://developer.github.com/v3/activity/events/types/#gollumevent
-type Page struct {
-	Name   string `json:"page_name"`
-	Title  string `json:"title"`
-	Action string `json:"action"`
-	Sha    string `json:"sha"`
-	URL    string `json:"html_url"`
-}
-
-// Sender represents the author of the commit
-type Sender struct {
-	Login     string `json:"login"`
-	AvatarURL string `json:"avatar_url"`
-}
-
-// GollumEvent houses all of the Page structs. For more detail see https://developer.github.com/v3/activity/events/types/#gollumevent
-type GollumEvent struct {
-	Pages  []Page `json:"pages"`
-	Sender Sender `json:"sender"`
-}
 
 // main is the handler for the action
 // Environment variables used are defined here: https://help.github.com/en/actions/configuring-and-managing-workflows/using-environment-variables#default-environment-variables
@@ -45,7 +25,8 @@ func main() {
 		os.Exit(1)
 	}
 
-	err = SendSlackMessage(event)
+	commsStrategy := notifier.Notifier{Strategy: &notifier.Slack{}}
+	err = commsStrategy.Strategy.Send(event)
 	if err != nil {
 		fmt.Println(msg)
 		os.Exit(1)
@@ -70,51 +51,16 @@ func ValidateConfiguration() (bool, string) {
 }
 
 // GetGollumEvent will unmarshal the JSON we receive from GitHub
-func GetGollumEvent() (*GollumEvent, error) {
+func GetGollumEvent() (*github.GollumEvent, error) {
 	file, err := ioutil.ReadFile(os.Getenv("GITHUB_EVENT_PATH"))
 	if err != nil {
 		return nil, fmt.Errorf("Unable to read the file defined GITHUB_EVENT_PATH, cannot carry on")
 	}
 
-	var gollum GollumEvent
+	var gollum github.GollumEvent
 	if err := json.Unmarshal([]byte(file), &gollum); err != nil {
 		return nil, fmt.Errorf("Unable to understand the JSON defined in GITHUB_EVENT_PATH, cannot carry on")
 	}
 
 	return &gollum, nil
-}
-
-// SendSlackMessage will send the required message to Slack.
-func SendSlackMessage(event *GollumEvent) error {
-	content := ""
-	for _, page := range event.Pages {
-		content += fmt.Sprintf("<%s|%s>\n", page.URL, page.Title)
-	}
-
-	attachments := []slack.Attachment{slack.Attachment{
-		Color:      "#2e5685",
-		Text:       content,
-		AuthorName: event.Sender.Login,
-		AuthorIcon: event.Sender.AvatarURL,
-	}}
-
-	msg := &slack.WebhookMessage{
-		Text:        "The following pages have changed in the wiki",
-		Attachments: attachments,
-	}
-
-	if os.Getenv("SLACK_USERNAME") != "" {
-		msg.Username = os.Getenv("SLACK_USERNAME")
-	}
-
-	if os.Getenv("SLACK_CHANNEL") != "" {
-		msg.Channel = os.Getenv("SLACK_CHANNEL")
-	}
-
-	if os.Getenv("DEBUG") != "" {
-		fmt.Println(msg)
-		return nil
-	}
-
-	return slack.PostWebhook(os.Getenv("SLACK_WEBHOOK"), msg)
 }
